moving some functions
